requirements.txt
acme==1.1.0
asgiref==3.5.0
astroid==2.11.3
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
blinker==1.4
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
dbus-python==1.2.16
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.3
filelock==3.6.0
future==0.18.2
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mock==3.0.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.2.0
packaging==21.3
parsedatetime==2.4
parso==0.8.3
pbr==5.4.5
pexpect==4.6.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==2.3.0
pylint==2.13.7
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sos==4.3
sqlparse==0.4.2
ssh-import-id==5.10
stripe==2.72.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
Twisted==18.9.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.7
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.14.1
wadllib==1.3.3
wrapt==1.14.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
